imp(archive): Use a catch all route for the archive section
